requirements.txt
flask
pandas
cython
numpy
scikit-surprise
chromadb